% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle2.R
\name{angle2}
\alias{angle2}
\title{angle2}
\usage{
angle2(p1, p2)
}
\arguments{
\item{p1}{Numeric matrix}

\item{p2}{Numeric matrix}
}
\value{
Named numeric vector of length equal to the number of rows in p1
}
\description{
Anti-clockwise angle from p1 as origin to p2 (x,y pairs matrices). The angle goes from 0 to 360.
}
