% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle2s.R
\name{angle2s}
\alias{angle2s}
\title{Compute Shortest angle anti-clockwise from p1 as origin to p2 (> -180 to 180)}
\usage{
angle2s(p1, p2)
}
\arguments{
\item{p1}{Numeric matrix}

\item{p2}{Numeric matrix}
}
\value{
Numeric vector of length equal to the number of rows in p1
}
\description{
Compute Shortest angle anti-clockwise from p1 as origin to p2 (> -180 to 180)
}
