% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/geometry.R,
%   R/geometry_wrappers.R
\name{angle2s_rcpp}
\alias{angle2s_rcpp}
\alias{angle2_rcpp}
\alias{Dn_rcpp}
\alias{angle2s_r}
\alias{angle2_r}
\alias{Dn_r}
\alias{angle2s}
\alias{angle2}
\alias{Dn}
\title{Anti-clockwise Angle}
\usage{
angle2s_rcpp(p1, p2)

angle2_rcpp(p1, p2)

Dn_rcpp(p_n, P_n)

angle2s_r(p1, p2)

angle2_r(p1, p2)

Dn_r(p_n, P_n)

angle2s(..., use = "cpp")

angle2(..., use = "cpp")

Dn(..., use = "cpp")
}
\arguments{
\item{p1, p2, p_n, P_n}{Numeric matrices with shape Nx2 (x and y).}

\item{...}{Arguments passed to the function implementation.}

\item{use}{Whether the R or C++ implementation is called.}
}
\value{
Named numeric vector of length equal to the number of rows in \code{p2}
or \code{p_n}.
}
\description{
Compute the anti-clockwise angles between two matrices with
the first column in each matrix containing x- and the second column
y-coordinates.
}
\details{
For \code{angle2s} and \code{angle2}, the angle is calculated from \code{p1} as origin to
\code{p2}. Therefore, \code{p1} must be a 1x2 matrix.

\code{Dn} calculates the angle between rows with the same index in
between $>0$ and 360 degrees. \code{angle2s} computes the angle in between
$>-180$ to 180 degrees and \code{angle2} in between $>0$ and 360 degrees.
}
