% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aTOd_rcpp}
\alias{aTOd_rcpp}
\title{aTOd}
\usage{
aTOd_rcpp(a)
}
\arguments{
\item{a}{Numeric vector - angles in degrees between 0 and 360}
}
\value{
Numeric Matrix of x and y coordinates (between -1 and 1) that are the signed (i.e., +/-) normalised difference between the xy points that generated the angle a
}
\description{
Compute sine and cosine of an angle.
}
