% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baUtility.R
\name{baUtility}
\alias{baUtility}
\title{Blocked Angle Utility}
\usage{
baUtility(p, BA)
}
\arguments{
\item{p}{Numeric vector with subject parameters.}

\item{BA}{Numeric vector of distances from each cell to closest pedestrian.}
}
\value{
Numeric vector with blocked angle utilities for each cell.
}
\description{
Blocked Angle Utility
}
