% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R, R/utility_wrappers.R
\name{baUtility_r}
\alias{baUtility_r}
\alias{baUtility}
\title{Blocked Angle Utility}
\usage{
baUtility_r(p, BA)

baUtility(p, BA, use = "cpp")
}
\arguments{
\item{p}{Numeric vector with subject parameters.}

\item{BA}{Numeric vector of distances from each cell to closest pedestrian.}

\item{use}{Whether the R or C++ implementation is called.}
}
\value{
Numeric vector with blocked angle utilities for each cell.
}
\description{
Blocked Angle Utility
}
