% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/block.R, R/block_wrappers.R
\name{bodyObjectOK_rcpp}
\alias{bodyObjectOK_rcpp}
\alias{bodyObjectOK_r}
\alias{bodyObjectOK}
\title{Check Cells}
\usage{
bodyObjectOK_rcpp(r, centres, objects, ok)

bodyObjectOK_r(r, centres, objects, ok)

bodyObjectOK(..., use = "cpp")
}
\arguments{
\item{r}{Numeric scalar radius of the body.}

\item{centres}{Numeric matrix with x- and y-coordinates of cell centers.}

\item{objects}{List containing a list for each object. An object has
two length-two numeric vectors of x- and y-coordinates.}

\item{ok}{Logical vector indicating which cells have been labeled 'ok'.}

\item{...}{Arguments passed to the function implementation.}

\item{use}{Whether the R or C++ implementation is called.}
}
\value{
Logical matrix (11x3) indicating whether cells are 'ok' or whether
there is an overlap between the body at cell coordinates and any objects.
}
\description{
Checks if cells are 'ok' or if the body radius at the cell coordinates
overlaps with any objects.
}
