% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/block.R, R/block_wrappers.R
\name{bodyObjectOverlap_rcpp}
\alias{bodyObjectOverlap_rcpp}
\alias{bodyObjectOverlap_r}
\alias{bodyObjectOverlap}
\title{Body-object Overlap}
\usage{
bodyObjectOverlap_rcpp(oL, r, okCentres)

bodyObjectOverlap_r(oL, r, okCentres)

bodyObjectOverlap(..., use = "cpp")
}
\arguments{
\item{oL}{List with two matrices for the start and end points of the
constituent lines. Each matrix has two columns with x- and y-coordinates and
a row for each line.}

\item{r}{Numeric scalar radius of the body.}

\item{okCentres}{Numeric matrix with x- and y-coordinates of cell centers
that have been labeled 'ok'.}

\item{...}{Arguments passed to the function implementation.}

\item{use}{Whether the R or C++ implementation is called.}
}
\value{
Logical vector indicating whether there is an overlap between the
body at coordinates in \code{okCentres} and any object lines \code{oL}.
}
\description{
Checks if a body with radius \code{r} overlaps with different object lines at
cell centre coordinates \code{okCentres}. These are coordinates of cells that
have been labeled 'ok' previously.
}
