% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_vd_rcpp}
\alias{c_vd_rcpp}
\title{c_vd}
\usage{
c_vd_rcpp(cells, p1, v1, a1_double, vels, angles)
}
\arguments{
\item{cells}{Integer vector indicating the set of pedestrian's cells}

\item{p1}{Numeric vector of pedestrian's xy coordinates}

\item{v1}{Scalar value of current velocity}

\item{a1_double}{Float of the current angle}

\item{vels}{Numeric matrix of velocities}

\item{angles}{Numeric matrix (33x3) of possible directions expressed in grades from 0 to 360}
}
\value{
Numeric matrix (33x2) of xy coordinates for each cell
}
\description{
Calculate cell centres for set of cells (index 1..33) for p1 heading at velocity v1 at angle a1 given time step tStep seconds.
}
