% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caUtility.R
\name{caUtility}
\alias{caUtility}
\title{Current Angle Utility}
\usage{
caUtility(p, angles = c(10, 20, 32.5, 50, 72.5)/90)
}
\arguments{
\item{p}{Numeric vector with subject parameters.}

\item{angles}{Numeric vector with angles of cones.}
}
\value{
Numeric vector with current angle utilities for each cell.
}
\description{
Current Angle Utility
}
