% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{caUtility_rcpp}
\alias{caUtility_rcpp}
\title{Current-angle Utility}
\usage{
caUtility_rcpp(
  aCA,
  bCA,
  bCAlr,
  angles = as.numeric(c(10, 20, 32.5, 50, 72.5)/90),
  n_rings = 3L
)
}
\arguments{
\item{aCA}{Numeric scalar power parameter.}

\item{bCA}{Numeric scalar scale parameter.}

\item{bCAlr}{Numeric scalar scale parameter.}

\item{angles}{Numeric vector with angles for which to calculate utility.}

\item{n_rings}{Integer number of cell rings for which to calculate utlity.}
}
\value{
Numeric vector of utilities for each cell.
}
\description{
Current-angle Utility
}
