% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/geometry.R,
%   R/geometry_wrappers.R
\name{coneNum_rcpp}
\alias{coneNum_rcpp}
\alias{coneNum_r}
\alias{coneNum}
\title{Cone Number}
\usage{
coneNum_rcpp(k)

coneNum_r(k)

coneNum(..., use = "cpp")
}
\arguments{
\item{k}{Numeric vector between 1 and 33.}
}
\value{
Numeric vector of length equal to k with cone numbers between
1 and 11.
}
\description{
Cone Number
}
