% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_rcpp_states}
\alias{create_rcpp_states}
\title{Rcpp States}
\usage{
create_rcpp_states(states)
}
\arguments{
\item{states}{List of variables representing a
state in the simulation.}
}
\value{
List of subjects states.
}
\description{
Helper function that transforms a state list simulated by predped into the
format required to do likelihood estimation.
}
