% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_rcpp_trace}
\alias{create_rcpp_trace}
\title{Rcpp Trace}
\usage{
create_rcpp_trace(trace)
}
\arguments{
\item{trace}{List of states in the simulation.}
}
\value{
List of simulation iterations containing subject states.
}
\description{
Helper function that transforms a trace list simulated by predped into the
format required to do likelihood estimation.
}
