% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/utility_extra.R,
%   R/utility_extra_wrappers.R
\name{destinationAngle_rcpp}
\alias{destinationAngle_rcpp}
\alias{destinationAngle_r}
\alias{destinationAngle}
\title{Angle to Destination}
\usage{
destinationAngle_rcpp(a, p1, P1)

destinationAngle_r(
  a,
  p1,
  P1,
  angles = c(72.5, 50, 32.5, 20, 10, 0, 350, 340, 327.5, 310, 287.5)
)

destinationAngle(..., use = "cpp")
}
\arguments{
\item{p1}{Numeric matrix with shape 1x2 (x and y).}

\item{P1}{Numeric matrix with shape Nx2 (x and y).}

\item{...}{Arguments passed to the function implementation.}

\item{use}{Whether the R or C++ implementation is called.}
}
\value{
Named numeric vector of length equal to the number of
rows N in \code{P1}.
}
\description{
Compute the angles between point \code{p1} and destinations \code{P1}.
}
\details{
The matrices \code{p1} and \code{p2} must have the same shape.
}
