% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist1_rcpp}
\alias{dist1_rcpp}
\title{dist1_rcpp}
\usage{
dist1_rcpp(p1, p2)
}
\arguments{
\item{p1}{Numeric matrix of a single point (i.e., 1 row, 2 xy columns)}

\item{p2}{Numeric matrix of multiple xy points}
}
\value{
Named numeric vector of length equal to the number of rows in p2
}
\description{
Compute distance from p1 to p2
}
