% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/geometry.R,
%   R/geometry_wrappers.R
\name{dist_rcpp}
\alias{dist_rcpp}
\alias{dist_r}
\alias{dist}
\title{Matrix-to-matrix Distance}
\usage{
dist_rcpp(p1, p2)

dist_r(p1, p2)

dist(..., use = "cpp")
}
\arguments{
\item{p1}{Numeric matrix with shape Nx2 (x and y).}

\item{p2}{Numeric matrix with shape Nx2 (x and y).}
}
\value{
Named numeric vector of length equal to the number of
rows N in \code{p1}.
}
\description{
Compute the Euclidian distance between two Nx2 matrices with the first
column in each matrix containing x- and the second column y-coordinates.
}
\details{
The matrices \code{p1} and \code{p2} must have the same shape.
}
