% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dn.R
\name{Dn}
\alias{Dn}
\title{Dn}
\usage{
Dn(p_n, P_n)
}
\arguments{
\item{p_n}{Numeric matrix of x and y coordinates}

\item{P_n}{Numeric matrix of x and y coordinates}
}
\value{
Named numeric vector of length equal to the number of rows in p1
}
\description{
Anti-clockwise angle to destination for all pedestrians. The angle goes from 0 to 360
}
