% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/utility_extra.R,
%   R/utility_extra_wrappers.R
\name{eObjects_rcpp}
\alias{eObjects_rcpp}
\alias{eObjects_r}
\alias{eObjects}
\title{Egocentric Objects}
\usage{
eObjects_rcpp(p1, p2, r)

eObjects_r(p1, p2, r)

eObjects(..., use = "cpp")
}
\arguments{
\item{p1}{Numeric matrix with shape 1x2 (x and y).}

\item{p2}{Numeric matrix with shape Nx2 (x and y).}

\item{r}{Numeric vector of length N.}

\item{...}{Arguments passed to the function implementation.}

\item{use}{Whether the R or C++ implementation is called.}
}
\value{
List:
\describe{
\item{ac}{Numeric matrix with rows as \emph{anticlockwise} end points to each row in \code{p2}
and columns as x- and y-coordinates.}
\item{cw}{Numeric matrix with rows as \emph{clockwise} end points to each row in \code{p2}
and columns as x- and y-coordinates.}
}
}
\description{
Compute the end points of line segments that are orthogonal to a line
between point \code{p1} and points \code{p2}. The width of the segments are equal to
the width \code{r} of \code{p2} from the perspective of \code{p1}.
}
