% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flUtility.R
\name{flUtility}
\alias{flUtility}
\title{Follow-leader Utility}
\usage{
flUtility(p, FL)
}
\arguments{
\item{p}{Numeric vector with subject parameters.}

\item{FL}{List of two numeric matrices:
\describe{
  \item{leaders}{Matrix with columns per leader and rows of their normalized angle disagreement and in-group status.}
  \item{dists}{Matrix with rows per leader and columns per cell with distances from each cell to chosen cell.}
}}
}
\value{
Numeric vector with follow-leader utilities for each cell.
}
\description{
Follow-leader Utility
}
