% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{flUtility_rcpp}
\alias{flUtility_rcpp}
\title{Compute flUtility}
\usage{
flUtility_rcpp(aFL, bFL, dFL, leaders, dists)
}
\arguments{
\item{aFL}{numeric vector}

\item{bFL}{numeric vector}

\item{dFL}{numeric vector}

\item{leaders}{named numeric matrix}
}
\value{
numeric vector of 33 elements
}
\description{
Compute flUtility
}
\examples{
flUtility_rcpp(1.5, 1, 4, leaders, dists)
}
