% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{free_cells_rcpp}
\alias{free_cells_rcpp}
\title{Check Cell Movement}
\usage{
free_cells_rcpp(agent, background, centers)
}
\arguments{
\item{agent}{Object of class \code{agent}.}

\item{background}{Object class \code{background}.}

\item{centers}{Numeric matrix with x- and y-coordinates of cell centers.}
}
\value{
Logical matrix (11x3) indicating whether cells are in the background,
and if there is a line of sight between the agent agent's position and the
cell center.
}
\description{
Checks if cells are outside the background, and if there is
a line of sight between the agent's position and the cell center.
}
