% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaUtility.R
\name{gaUtility}
\alias{gaUtility}
\title{Goal Angle Utility}
\usage{
gaUtility(p, GA)
}
\arguments{
\item{p}{Numeric vector with subject parameters.}

\item{GA}{Numeric vector with angles to next goal.}
}
\value{
Numeric vector with goal angle utilities for each cell.
}
\description{
Goal Angle Utility
}
