% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R, R/utility_wrappers.R
\name{gaUtility_r}
\alias{gaUtility_r}
\alias{gaUtility}
\title{Goal Angle Utility}
\usage{
gaUtility_r(p, GA)

gaUtility(p, GA, use = "cpp")
}
\arguments{
\item{p}{Numeric vector with subject parameters.}

\item{GA}{Numeric vector with angles to next goal.}

\item{use}{Whether the R or C++ implementation is called.}
}
\value{
Numeric vector with goal angle utilities for each cell.
}
\description{
Goal Angle Utility
}
