% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msumlogLike_rcpp.R
\name{get_cell_nest}
\alias{get_cell_nest}
\title{Cell Nest Association}
\usage{
get_cell_nest()
}
\value{
Integer matrix indicating which cells belong to which nests. Indices start at zero to match C++ indexing.
}
\description{
Cell Nest Association
}
