% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mum}
\alias{get_mum}
\title{Nest Association}
\usage{
get_mum(p)
}
\arguments{
\item{p}{Numeric vector with subject parameters.}
}
\value{
Numeric vector with nest association parameters.
}
\description{
Nest Association
}
