% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/utility_extra.R,
%   R/utility_extra_wrappers.R
\name{getBuddy_rcpp}
\alias{getBuddy_rcpp}
\alias{getBuddy_r}
\alias{getBuddy}
\title{Buddies}
\usage{
getBuddy_rcpp(n, group, a, p_pred, centres, objects, pickBest, state)

getBuddy_r(n, group, a, p_pred, centres, objects, pickBest = FALSE, state)

getBuddy(
  n,
  group,
  a,
  p_pred,
  centres,
  objects,
  pickBest = FALSE,
  state,
  use = "cpp"
)
}
\arguments{
\item{n}{Integer scalar index of current pedestrian in pedestrian matrix.}

\item{group}{Integer vector of group memberships.}

\item{a}{Numeric vector of heading angles for each pedestrian.}

\item{p_pred}{Numeric matrix with shape Nx2 (x and y) as predicted positions
of all pedestrians.}

\item{centres}{Numeric matrix with x and y for each cell centre (33x2).}

\item{objects}{List containing a list for each object. An object has
two length-two numeric vectors of x- and y-coordinates.}

\item{pickBest}{Logical salar indicating if all or the best buddy should be
returned.}

\item{state}{List with state data.}

\item{use}{Whether the R or C++ implementation is called.}
}
\value{
List:
\describe{
\item{leaders}{Numeric matrix with two rows for every buddy
(columns). The first row is the cell of the leader, the second the
difference in heading angle between pedestrian \code{n} and buddy.}
\item{dists}{Numeric matrix with distances for each cell (columns) to
each buddy (rows).}
}
}
\description{
Compute the distance from each cell to buddies.
}
\details{
Returns \code{NULL} if no other pedestrians are present, if they cannot be seen,
if they are not in front, or if no other pedestrian has the same group as
pedestrian \code{n}.
}
