% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headingAngle.R
\name{headingAngle}
\alias{headingAngle}
\title{headingAngle}
\usage{
headingAngle(
  a2,
  a1,
  angles = c(72.5, 50, 32.5, 20, 10, 0, 350, 340, 327.5, 310, 287.5)
)
}
\arguments{
\item{a2}{Numeric vector}

\item{a1}{Numeric vector}
}
\value{
Numeric matrix whose rows and columns are the same length of a2 and a1
}
\description{
Absolute angular difference between 11 directions with zero cone having angle a1 and a2
}
