% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{headingAngle_rcpp}
\alias{headingAngle_rcpp}
\title{headingAngle}
\usage{
headingAngle_rcpp(a2, a1)
}
\arguments{
\item{a2}{Numeric vector}

\item{a1}{Numeric vector}
}
\value{
Numeric matrix whose rows and columns are the same length of a2 and a1
}
\description{
Absolute angular difference between 11 directions with zero cone having angle a1 and a2
}
