% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Iangle.R
\name{Iangle}
\alias{Iangle}
\title{Iangle}
\usage{
Iangle(
  p1,
  a1,
  p2,
  border = c(-85, -60, -40, -25, -15, -5, 5, 15, 25, 40, 60, 85)
)
}
\arguments{
\item{p1}{Numeric matrix}

\item{a1}{Numeric vector - angles in degrees between 0 and 360}

\item{p2}{Numeric matrix}
}
\value{
Numeric Vector of indices whereby the angle is within border's bins
}
\description{
Which angle cone (1..11, NA means outside of view) is p2 in relative to p1 heading at angel a1
}
