% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/geometry.R,
%   R/geometry_wrappers.R
\name{Iangle_rcpp}
\alias{Iangle_rcpp}
\alias{Iangle_r}
\alias{Iangle}
\title{Bin Angles between Matrices}
\usage{
Iangle_rcpp(p1, a1, p2)

Iangle_r(
  p1,
  a1,
  p2,
  border = c(-85, -60, -40, -25, -15, -5, 5, 15, 25, 40, 60, 85)
)

Iangle(..., use = "cpp")
}
\arguments{
\item{p1}{Numeric matrix with shape Nx2 (x and y).}

\item{a1}{Numeric vector containing angles in degrees between 0 and 360.}

\item{p2}{Numeric matrix with shape Nx2 (x and y).}
}
\value{
Numeric vector of bin indices.
}
\description{
Compute anti-clockwise angles between two Nx2 matrices with the first
column in each matrix containing x- and the second column y-coordinates
(> 0 to 360 degrees) using \link[=dest]{angle2} in relation to a scalar
angle. Bin relative angles according to a numeric vector of breaks.
}
\details{
Returns \code{NA} if relative angle is \code{NA} or outside the breaks. Excludes the
left and includes the right border of every bin.
}
