% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/utility_extra.R,
%   R/utility_extra_wrappers.R
\name{iCones2Cells_rcpp}
\alias{iCones2Cells_rcpp}
\alias{iCones2Cells_r}
\alias{iCones2Cells}
\title{Transform Intersecting Cones to Cells}
\usage{
iCones2Cells_rcpp(iC, v, tStep = 0.5)

iCones2Cells_r(iC, v, vels = c(1.5, 1, 0.5))

iCones2Cells(..., use = "cpp")
}
\arguments{
\item{iC}{Numeric vector with intersecting cones returned by
\link[=iCones]{iCones}.}

\item{v}{Numeric scalar velocity of current pedestrian.}

\item{tStep}{Numeric scalar velocity scaling factor.}

\item{vels}{Numeric vector with velocities for each ring. Must be length 3.}

\item{...}{Arguments passed to the function implementation.}

\item{use}{Whether the R or C++ implementation is called.}
}
\value{
Numeric vector with distances.
}
\description{
Transform intersecting cones into vector with cell names containing
distances to each cell ring for a pedestrian moving with speed \code{v}.
}
