% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/utility_extra.R,
%   R/utility_extra_wrappers.R
\name{iCones_rcpp}
\alias{iCones_rcpp}
\alias{iCones_r}
\alias{iCones}
\title{Intersecting Cones}
\usage{
iCones_rcpp(p1, a, p2, r, objects)

iCones_r(p1, a, p2, r, objects)

iCones(..., use = "cpp")
}
\arguments{
\item{p1}{Numeric matrix with shape 1x2 (x and y).}

\item{p2}{Numeric matrix with shape Nx2 (x and y) indicating the positions
of all \emph{other} pedestrians.}

\item{r}{Numeric vector with radii of pedestrian bodies.}

\item{objects}{List containing a list for each object. An object has
two length-two numeric vectors of x- and y-coordinates.}

\item{...}{Arguments passed to the function implementation.}

\item{use}{Whether the R or C++ implementation is called.}

\item{a1}{Numeric scalar heading angle of pedestrian \code{p1}.}
}
\value{
Named numeric vector with intersecting cones or NULL. Names
indicate the cone number and the elements the distances to the intersection.
}
\description{
Compute intersecting cones between pedestrian \code{p1} closest pedestrians \code{p2}
from the perspective of \code{p1}.
}
\details{
Returns \code{NULL} if no other pedestrians are present, if they cannot be seen,
if they are not in front of pedestrian \code{n}, or if their cones are not
intersecting.
}
