% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idUtility.R
\name{idUtility}
\alias{idUtility}
\title{Interpersonal Distance Utility}
\usage{
idUtility(p, n, ID, ok, group)
}
\arguments{
\item{p}{Numeric vector with subject parameters.}

\item{n}{Integer scalar indexing the subject in the state.}

\item{ID}{Numeric matrix of predicted distances from the subject to other pedestrians in the front.}

\item{ok}{Logical matrix indicating if cells are blocked.}

\item{group}{Integer vector with group indices for each pedestrian.}
}
\value{
Numeric vector with interpersonal distance utilities for each cell.
}
\description{
Interpersonal Distance Utility
}
