% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R, R/utility_wrappers.R
\name{idUtility_r}
\alias{idUtility_r}
\alias{idUtility}
\title{Interpersonal Distance Utility}
\usage{
idUtility_r(p, n, ID, ok, group)

idUtility(p, n, ID, ok, group, use = "cpp")
}
\arguments{
\item{p}{Numeric vector with subject parameters.}

\item{n}{Integer scalar indexing the subject in the state.}

\item{ID}{Numeric matrix of predicted distances from the subject to other pedestrians in the front.}

\item{ok}{Logical matrix indicating if cells are blocked.}

\item{group}{Integer vector with group indices for each pedestrian.}

\item{use}{Whether the R or C++ implementation is called.}
}
\value{
Numeric vector with interpersonal distance utilities for each cell.
}
\description{
Interpersonal Distance Utility
}
