% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{idUtility_rcpp}
\alias{idUtility_rcpp}
\title{Compute IDUtility}
\usage{
idUtility_rcpp(bID, dID, aID, n, ok, group, ID_ = NULL)
}
\arguments{
\item{bID}{Numeric scalar.}

\item{dID}{Numeric scalar.}

\item{aID}{Numeric scalar.}

\item{n}{Numeric scalar indexing the subject in the state.}

\item{ok}{Logical matrix indicating if cells are blocked.}

\item{group}{Named numeric scalar with group indices for each pedestrian.}

\item{ID_}{Numeric matrix of the type NULL - if not NULL, ID_ is a Numeric matrix of predicted distances from the subject to other pedestrians in the front.}
}
\value{
Numeric vector with interpersonal distance utilities for each cell.
}
\description{
Inter-personal distance utility for cell 1..33. b parameter divided by
sum over power of distances between bodies for cell to all inFront peds.
}
