% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msumlogLike_rcpp.R
\name{like_state_rcpp}
\alias{like_state_rcpp}
\title{State Likelihood}
\usage{
like_state_rcpp(state, p, n, nests, alpha, cell_nest)
}
\arguments{
\item{state}{List of state variables.}

\item{p}{Numeric vector of subject parameters.}

\item{n}{Integer scalar indexing the subject in the state.}

\item{nests}{List of vectors with utility indices.}

\item{alpha}{List of vectors with alpha values.}

\item{cell_nest}{Integer matrix associating the cells with the nests.}
}
\value{
Probability of the state given the parameters.
}
\description{
State Likelihood
}
