% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msumloglike_rcpp.R
\name{like_states_rcpp}
\alias{like_states_rcpp}
\title{Trace Likelihood}
\usage{
like_states_rcpp(trace, p, nests, alpha, cell_nest)
}
\arguments{
\item{trace}{List of lists with state variables.}

\item{p}{Numeric matrix with subject parameters.}

\item{nests}{List of vectors with utility indices.}

\item{alpha}{List of vectors with alpha values.}

\item{cell_nest}{Integer matrix associating the cells with the nests.}
}
\value{
List with a numeric vector for each state containing the state likelihoods.
}
\description{
Trace Likelihood
}
