% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{like_subject}
\alias{like_subject}
\title{Subject Log-likelihood}
\usage{
like_subject(subject, p, n, nests, alpha, cell_nest, min_like = 1e-10)
}
\arguments{
\item{subject}{List with subject data.}

\item{p}{Numeric vector with subject parameters.}

\item{n}{Integer scalar subject index.}

\item{nests}{List of vectors with utility indices.}

\item{alpha}{List of vectors with alpha values.}

\item{cell_nest}{Numeric matrix with nest indices for each cell.}

\item{min_like}{Numeric scalar minimum likelihood return value.}
}
\value{
Numeric scalar subject log likelihood.
}
\description{
Calculate the log-likelihood of an observation for a single pedestrian.
}
