% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see.R
\name{line.line.intersection}
\alias{line.line.intersection}
\title{intersections inside L1 (from P1 to P2) and L2 (from P3 to P4).}
\source{
Weisstein, Eric W. "Line-Line Intersection."
From MathWorld--A Wolfram Web Resource.
\url{http://mathworld.wolfram.com/Line-LineIntersection.html}
}
\usage{
line.line.intersection(P1, P2, P3, P4, interior.only = FALSE)
}
\value{
Vector containing x,y coordinates of intersection of L1
and L2.  If L1 and L2 are parallel, this is infinite-valued.  If
\code{interior.only} is \code{TRUE}, then when the intersection
does not occur between P1 and P2 and P3 and P4, a vector
containing \code{NA}s is returned.
}
\description{
intersections inside L1 (from P1 to P2) and L2 (from P3 to P4).
}
\examples{
## Intersection of two lines
line.line.intersection(c(0, 0), c(1, 1), c(0, 1), c(1, 0))

## Two lines that don't intersect
line.line.intersection(c(0, 0), c(0, 1), c(1, 0), c(1, 1))
}
\author{
David Sterratt
}
