% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{line_line_intersection_rcpp}
\alias{line_line_intersection_rcpp}
\title{Two-line Intersection}
\source{
Weisstein, Eric W. "Line-Line Intersection.
"From MathWorld--A Wolfram Web Resource.
\url{http://mathworld.wolfram.com/Line-LineIntersection.html}
}
\usage{
line_line_intersection_rcpp(P1, P2, P3, P4, interior_only = FALSE)
}
\arguments{
\item{P1, P2, P3, P4}{Numeric vector with x- and y-coordinates.}
}
\value{
A numeric vector with x- and y-coordinates defining the
intersecting point between the two lines.
}
\description{
Calculate the intersection between two lines.
}
\details{
The first line is defined by \code{P1} and \code{P1}, the second by
\code{P3} \code{P4}. Returns \code{Inf} when the lines are parallel. If
\code{interior.only} is \code{TRUE}, returns \code{NA} if the intersection
is not within the plane spanning between the two lines.
}
\author{
David Sterratt
}
