\name{m4ma-package}
\alias{m4ma-package}
\alias{m4ma}
\docType{package}
\title{
  Implementation of the predictive pedestrian agent model
}
\description{
  The Minds for Mobile Agents (M4MA) project
    aims to realistically model how pedestrians with diverse individual
    characteristics move around spatial and social environments while
    fulfilling complex goals. M4MA’s “predictive pedestrian” agent model
    simulates and explains the causes of pedestrian movement by
    combining tools from cognitive decision modelling and behavioural 
    data science.
}
\details{
  
}
\author{
Malte Luken <m.luken@esciencecenter.nl> and Eva Viviani <e.viviani@esciencecenter.nl>

Maintainer: Charlotte Tanis <c.c.tanis@uva.nl>
}
\references{
  
}
\keyword{ package }
\seealso{
  
}
\examples{
  \dontrun{
     
  }
}
