% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{minAngle_rcpp}
\alias{minAngle_rcpp}
\title{minAngle}
\usage{
minAngle_rcpp(a1_double, a2)
}
\arguments{
\item{a1_double}{scalar vector}

\item{a2}{Numeric vector}
}
\value{
Numeric vector of length equal to a1
}
\description{
Shortest absolute angle between a1 and a2
}
