% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{msumlogLike}
\alias{msumlogLike}
\title{Trace Log-likelihood}
\usage{
msumlogLike(p, trace, nests, alpha, cell_nest, min_like = 1e-10, mult = -1)
}
\arguments{
\item{p}{Numeric matrix with subject parameters for each subject.}

\item{trace}{List of lists of lists with subject data.}

\item{nests}{List of vectors with utility indices.}

\item{alpha}{List of vectors with alpha values.}

\item{cell_nest}{Numeric matrix with nest indices for each cell.}

\item{min_like}{Numeric scalar minimum likelihood return value.}

\item{mult}{Numeric scalar likelihood sum multiplier.}
}
\value{
Numeric scalar trace log-likelihood.
}
\description{
Calculate the log-likelihood of a trace of states as the sum over states and subject log-likelihoods.
}
