% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msumlogLike_rcpp.R
\name{msumlogLike_rcpp}
\alias{msumlogLike_rcpp}
\title{Log-likelihood Sum}
\usage{
msumlogLike_rcpp(trace, p, minLike = 1e-10, mult = -1)
}
\arguments{
\item{trace}{List of lists with state variables.}

\item{p}{Numeric matrix with subject parameters.}

\item{minLike}{Numeric scalar indicating minimum likelihood value.}

\item{mult}{Numeric scalar indicating likelihood sum multiplicator.}
}
\value{
Numeric scalar indicating log-likelihood sum of trace given parameters.
}
\description{
Log-likelihood Sum
}
