% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/block.R, R/block_wrappers.R
\name{object2lines_rcpp}
\alias{object2lines_rcpp}
\alias{object2lines_r}
\alias{object2lines}
\title{Rectangle to Lines}
\usage{
object2lines_rcpp(o)

object2lines_r(o)

object2lines(..., use = "cpp")
}
\arguments{
\item{o}{List with length-two numeric vectors of x- and y-coordinates of an
object.}

\item{...}{Arguments passed to the function implementation.}

\item{use}{Whether the R or C++ implementation is called.}
}
\value{
List with two matrices for the start and end points of the
constituent lines. Each matrix has two columns with x- and y-coordinates and
a row for each line.
}
\description{
Transforms a list of objects into a list of their constituent lines.
}
