% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCNL.R, R/utility_wrappers.R
\name{pCNL_r}
\alias{pCNL_r}
\alias{pCNL}
\title{Probability of the Conditional Nested Logit Model}
\usage{
pCNL_r(cell, V, muM = rep(1, length(nests)), nests, alpha, mu, cellNest)

pCNL(
  cell,
  V,
  muM = rep(1, length(nests)),
  nests,
  alpha,
  mu,
  cellNest,
  use = "cpp"
)
}
\arguments{
\item{cell}{Alternative vector with nest indices.}

\item{V}{Vector with utility for each alternative.}

\item{muM}{Vector with nests associations ranging between 0 and 1.}

\item{nests}{List of vectors with utility indices.}

\item{alpha}{List of vectors with alpha values.}

\item{mu}{General nest association ranging between 0 and 1.}

\item{cellNest}{Integer matrix associating the cells with the nests.}

\item{use}{Whether the R or C++ implementation is called.}
}
\value{
Probability of alternative \code{cell} given \code{V}, \code{alpha}, and parameters \code{muM} and \code{mu}.
}
\description{
Probability of the Conditional Nested Logit Model
}
