% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcnl_rcpp}
\alias{pcnl_rcpp}
\title{Probability of the Conditional Nested Logit Model}
\usage{
pcnl_rcpp(cell, u, mum, nests, alpha, mu)
}
\arguments{
\item{cell}{Alternative vector with nest indices.}

\item{u}{Vector with utility for each alternative.}

\item{mum}{Vector with nests associations ranging between 0 and 1.}

\item{nests}{List of vectors with utility indices.}

\item{alpha}{List of vectors with alpha values.}

\item{mu}{General nest association ranging between 0 and 1.}
}
\value{
Probability of alternative \code{cell} given \code{u}, \code{alpha}, and parameters \code{mum} and \code{mu}.
}
\description{
Probability of the Conditional Nested Logit Model
}
