% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pmnl_rcpp}
\alias{pmnl_rcpp}
\title{Probability of the Multinomial Logit Model}
\usage{
pmnl_rcpp(cell, utility, ok)
}
\arguments{
\item{cell}{Integer with the index of the cell to compute the probability for.}

\item{utility}{Numeric vector with the utility for each cell.}

\item{ok}{Logical matrix with the same shape as utility indicating which cells are available.}
}
\value{
Probability of alternative \code{cell} given \code{utility} and \code{ok}.
}
\description{
DEPRECATED. Computed the probability of the mulinomial logit model given a cell and utility vector.
}
