% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/utility_extra.R,
%   R/utility_extra_wrappers.R
\name{predClose_rcpp}
\alias{predClose_rcpp}
\alias{predClose_r}
\alias{predClose}
\title{Predicted Distance to Close Front}
\usage{
predClose_rcpp(n, p1, a1, p2, r, centres, p_pred, objects)

predClose_r(n, p1, a1, p2, r, centres, p_pred, objects)

predClose(..., use = "cpp")
}
\arguments{
\item{n}{Integer scalar index of current pedestrian in pedestrian matrix.}

\item{p1}{Numeric matrix with shape 1x2 (x and y) indicating the position
of pedestrian \code{n}.}

\item{a1}{Numeric scalar heading angle of pedestrian \code{n}.}

\item{p2}{Numeric matrix with shape Nx2 (x and y) indicating the positions
of all pedestrians.}

\item{r}{Numeric vector with radii of pedestrian bodies.}

\item{centres}{Numeric matrix with x and y for each cell centre (33x2).}

\item{p_pred}{Numeric matrix with shape Nx2 (x and y) as predicted positions
of all pedestrians.}

\item{objects}{List containing a list for each object. An object has
two length-two numeric vectors of x- and y-coordinates.}

\item{...}{Arguments passed to the function implementation.}

\item{use}{Whether the R or C++ implementation is called.}
}
\value{
Numeric matrix with rows for each other pedestrian and columns for
each cell (Nx33) or NULL.
}
\description{
Compute the distance from edge of body of pedestrian \code{n} to other pedestrians
in front that can be seen.
}
\details{
Returns \code{NULL} if no other pedestrians are present, if they cannot be seen,
or if they are not in front of pedestrian \code{n}.
}
