% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predped_env}
\alias{predped_env}
\title{Indicator of functions to call (R or C++)}
\usage{
predped_env

predped_env
}
\description{
Indicator of functions to call (R or C++)

Indicator of functions to call (R or C++)
}
\keyword{internal}
