% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psUtility.R
\name{psUtility}
\alias{psUtility}
\title{Preferred Speed Utility}
\usage{
psUtility(p, v, d)
}
\arguments{
\item{p}{Numeric vector with subject parameters.}

\item{v}{Numeric scalar indicating the current speed.}

\item{d}{Numeric scalar indicating the distance to next goal.}
}
\value{
Numeric vector with preferred speed utilities for each cell.
}
\description{
Preferred Speed Utility
}
