% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R, R/utility_wrappers.R
\name{psUtility_r}
\alias{psUtility_r}
\alias{psUtility}
\title{Preferred Speed Utility}
\usage{
psUtility_r(p, v, d)

psUtility(p, v, d, use = "cpp")
}
\arguments{
\item{p}{Numeric vector with subject parameters.}

\item{v}{Numeric scalar indicating the current speed.}

\item{d}{Numeric scalar indicating the distance to next goal.}

\item{use}{Whether the R or C++ implementation is called.}
}
\value{
Numeric vector with preferred speed utilities for each cell.
}
\description{
Preferred Speed Utility
}
