% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{psUtility_rcpp}
\alias{psUtility_rcpp}
\title{Compute psUtility}
\usage{
psUtility_rcpp(aPS, sPref, sSlow, bPS, v, d)
}
\arguments{
\item{aPS}{numeric vector}

\item{sPref}{numeric vector}

\item{sSlow}{numeric vector}

\item{bPS}{numeric vector}

\item{v}{numeric vector}

\item{d}{numeric vector}
}
\value{
a numeric vector of length equals to d's length
}
\description{
Compute psUtility
}
\examples{

psUtility_rcpp(aPS = 1, sPref = 3,  sSlow = 2, bPS = 1,  v = 1, d = 11.46)
}
