% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleVel.R
\name{scaleVel}
\alias{scaleVel}
\title{scaleVel}
\usage{
scaleVel(v, tStep = 0.5)
}
\arguments{
\item{v}{Numeric vector}

\item{tStep}{double}
}
\value{
Numeric vector of scaled velocity of same length of v
}
\description{
Scale velocity by time step (tStep)
}
