% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seesCurrentGoal_rcpp}
\alias{seesCurrentGoal_rcpp}
\title{Current Goal in Sight}
\usage{
seesCurrentGoal_rcpp(n, state, objects, offset = 0L)
}
\arguments{
\item{n}{Integer scalar subject index.}

\item{state}{List of list with state data.}

\item{objects}{List containing a list for each object. An object has
two length-two numeric vectors of x- and y-coordinates.}
}
\value{
\code{TRUE} if the goal is in sight, \code{FALSE} otherwise.
}
\description{
Checks whether the current goal in a \code{state} can be seen by
subject \code{n}, or if it is occluded by \code{objects}.
}
\examples{
objects = list(
  list(x = c(0.5, 0.5), y = c(0.5, 0.5))
)

state = list(
  p = matrix(c(0, 0), 1, 2),
  P = list(
    matrix(c(1, 1), 1, 2)
  )
)

seesCurrentGoal_rcpp(1, state, objects)
# FALSE

}
