% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/see.R, R/see_wrappers.R
\name{seesGoal_rcpp}
\alias{seesGoal_rcpp}
\alias{seesGoal_r}
\alias{seesGoal}
\title{Goal in Sight}
\usage{
seesGoal_rcpp(p_n, P_n, objects)

seesGoal_r(p_n, P_n, objects)

seesGoal(..., use = "cpp")
}
\arguments{
\item{p_n, P_n}{Numeric vector with x- and y-coordinates.}

\item{objects}{List containing a list for each object. An object has
two length-two numeric vectors of x- and y-coordinates.}

\item{...}{Arguments passed to the function implementation.}

\item{use}{Whether the R or C++ implementation is called.}
}
\value{
\code{TRUE} if the goal is in sight, \code{FALSE} otherwise.
}
\description{
Checks whether a goal \code{P} can be seen from point \code{p}, or if it is
occluded by \code{objects}.
}
\examples{
objects = list(
  list(x = c(0.5, 0.5), y = c(0.5, 0.5))
)

seesGoal_rcpp(c(0, 0), c(1, 1), objects)
# FALSE

}
