% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/see.R, R/see_wrappers.R
\name{seesGoalOK_rcpp}
\alias{seesGoalOK_rcpp}
\alias{seesGoalOK_r}
\alias{seesGoalOK}
\title{Goal in Sight from Cell Centres}
\usage{
seesGoalOK_rcpp(n, objects, state, centres, ok)

seesGoalOK_r(n, objects, state, centres, ok)

seesGoalOK(..., use = "cpp")
}
\arguments{
\item{n}{Integer scalar subject index.}

\item{objects}{List containing a list for each object. An object has
two length-two numeric vectors of x- and y-coordinates.}

\item{state}{List of list with state data.}

\item{centres}{Numeric matrix with 33 cell centres as rows and
x- and y-coordinates as columns.}

\item{ok}{Logical vector of length 33 indicating which cells are
marked as 'ok'.}

\item{...}{Arguments passed to the function implementation.}

\item{use}{Whether the R or C++ implementation is called.}
}
\value{
Logical vector of length 33 indicating from which 'ok' cells the
current goal can be seen.
}
\description{
Checks whether the current goal in a \code{state} can be seen by
subject \code{n} from cell \code{centres} that are marked as \code{ok},
or if it is occluded by \code{objects}.
}
\examples{
objects = list(
  list(x = c(0.5, 0.5), y = c(0.5, 0.5))
)

state = list(
  p = matrix(c(0, 0), 1, 2),
  P = list(
    matrix(c(1, 1), 1, 2)
  )
)

# Random centres and ok
set.seed(123)
centres = matrix(rnorm(66), 2, 2)

ok = as.logical(sample(c(0, 1), 33, replace = TRUE, prob = c(0.3, 0.7)))

seesGoalOK_rcpp(1, objects, state, centres, ok)

}
