% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seesMany_rcpp}
\alias{seesMany_rcpp}
\title{Multiple Goals in Sight}
\usage{
seesMany_rcpp(p1, ps, objects)
}
\arguments{
\item{p1}{Numeric vector with x- and y-coordinates.}

\item{ps}{Numeric Matrix with a row for every goal and x- and y-coordinates
as columns.}

\item{objects}{List containing a list for each object. An object has
two length-two numeric vectors of x- and y-coordinates.}
}
\value{
Logical vector with \code{TRUE} if a goal is in sight, \code{FALSE}
otherwise.
}
\description{
Checks which of the goals \code{ps} can be seen from point \code{p1},
or if they occluded by \code{objects}.
}
\examples{
objects = list(
  list(x = c(0.5, 0.5), y = c(0.5, 0.5))
)

goals = rbind(
  c(1, 0),
  c(1, 1)
)

seesGoal_rcpp(c(0, 0), goals, objects)
# TRUE FALSE

}
