% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{utility}
\alias{utility}
\title{Total Utility of Cells}
\usage{
utility(p, v, d, ba_, ga, id_, is_ingroup, id_pre_utility, fl_, wb_, ok)
}
\arguments{
\item{p}{Numeric vector of subject parameters.}

\item{v}{Numeric scalar indicating the current speed.}

\item{d}{Numeric scalar indicating the distance to next goal.}

\item{ba_}{NULL or numeric vector of distances from each cell to closest pedestrian.}

\item{ga}{Numeric vector of angles to next goal.}

\item{id_}{NULL or numeric matrix of predicted distances from the subject to other pedestrians in the front.}

\item{is_ingroup}{Logical vector indicating which subjects in \code{ID_} are part of the ingroup.}

\item{id_pre_utility}{Numeric pre-calculated interpersonal distance utility vector (contains \code{-Inf} for blocked cells and \code{0} otherwise).}

\item{fl_}{NULL or list of numeric matrices:
\describe{
\item{leaders}{Numeric matrix of buddy positions and angles.}
\item{dists}{Matrix with rows per leader and columns per cell with distances from each cell to chosen cell.}
}}

\item{wb_}{NULL or list of numeric matrices:
\describe{
\item{buddies}{Matrix with columns per leader and rows of their normalized angle disagreement and in-group status. # needs rewrite}
\item{dists}{Numeric matrix of distances from cells' centers to closest buddy. # needs rewrite}
}}

\item{ok}{Logical matrix indicating if cells are blocked.}
}
\value{
Numeric vector with total utility for each cell.
}
\description{
Total Utility of Cells
}
