% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_rcpp.R
\name{utility_rcpp}
\alias{utility_rcpp}
\title{Total Utility of Alternatives}
\usage{
utility_rcpp(p, n, v, d, ga, ba, id, fl, wb, ok, group)
}
\arguments{
\item{p}{Vector with subject parameters.}

\item{v}{Scalar indicating the current speed..}

\item{d}{Scalar indicating the distance to next goal.}

\item{ga}{Vector with angles to next goal.}
}
\value{
Vector with total utility for each alternative.
}
\description{
Total Utility of Alternatives
}
