% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbUtility.R
\name{wbUtility}
\alias{wbUtility}
\title{Walk-beside Utility}
\usage{
wbUtility(p, WB)
}
\arguments{
\item{p}{Numeric vector with subject parameters.}

\item{WB}{Numeric vector of distances from cells' centers to closest buddy.}
}
\value{
Numeric vector with walk-beside utilities for each cell.
}
\description{
Walk-beside Utility
}
