% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wbUtility_rcpp}
\alias{wbUtility_rcpp}
\title{Walk-beside Utility}
\usage{
wbUtility_rcpp(aWB, bWB, buddies, dists)
}
\arguments{
\item{aWB}{Numeric scalar power parameter.}

\item{bWB}{Numeric scalar scale parameter.}

\item{buddies}{Numeric matrix of buddy positions and angles. # needs rewrite}

\item{dists}{Numeric matrix of distances from cells' centers to closest buddy. # needs rewrite}
}
\value{
Numeric vector of utilities for each cell.
}
\description{
Walk-beside Utility
}
