% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_wrappers.R
\name{wrapper}
\alias{wrapper}
\title{Execute R or C++ Implementation}
\usage{
wrapper(name)
}
\arguments{
\item{name}{The name of the function to be wrapped.}
}
\value{
The wrapped function with the additional \code{use} argument to switch
implementations.
}
\description{
This is a wrapper function that takes a function name and returns a function
that calls the function with an additional argument \code{use} to switch between
an R and C++ version.
}
