% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/geometry.R,
%   R/geometry_wrappers.R
\name{aTOd_rcpp}
\alias{aTOd_rcpp}
\alias{aTOd_r}
\alias{aTOd}
\title{Sine and Cosine}
\usage{
aTOd_rcpp(a)

aTOd_r(a)

aTOd(..., use = "cpp")
}
\arguments{
\item{a}{Numeric vector containing angles in degrees between 0 and 360.}

\item{...}{Arguments passed to the function implementation.}

\item{use}{Whether the R or C++ implementation is called.}
}
\value{
Numeric matrix of x and y values (between -1 and 1) that are
the signed normalized differences between the xy-points leading to
the angles \code{a}.
}
\description{
Compute sine and cosine of angles.
}
