% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{baUtility_rcpp}
\alias{baUtility_rcpp}
\title{Blocked-angle Utility}
\usage{
baUtility_rcpp(aBA, bBA, BA, idx_BA)
}
\arguments{
\item{aBA}{Power parameter.}

\item{bBA}{Scale parameter.}

\item{BA}{Numeric vector of distances from cells to closest pedestrians.}

\item{idx_BA}{Integer vector of cell indices.}
}
\value{
Numeric Vector of utilities for each cell.
}
\description{
Blocked-angle Utility
}
