% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/geometry.R,
%   R/geometry_wrappers.R
\name{c_vd_rcpp}
\alias{c_vd_rcpp}
\alias{c_vd_r}
\alias{c_vd}
\title{Cell Centers}
\usage{
c_vd_rcpp(cells, p1, v1, a1, vels, angles, tStep = 0.5)

c_vd_r(
  cells,
  p1,
  v1,
  a1,
  vels = matrix(rep(c(1.5, 1, 0.5), each = 11), ncol = 3),
  angles = matrix(rep(c(72.5, 50, 32.5, 20, 10, 0, 350, 340, 327.5, 310, 287.5), times
    = 3), ncol = 3)
)

c_vd(...)
}
\arguments{
\item{cells}{Integer vector indicating cell indices between 1 and 33.}

\item{p1}{Numeric vector of length two (x and y).}

\item{v1}{Numeric scalar velocity.}

\item{a1}{Numeric scalar angle.}

\item{vels}{Numeric matrix (33x3) of velocities for each cell.}

\item{angles}{Numeric matrix (33x3) of angles for each cell.}

\item{tStep}{Numeric scalar velocity scaling factor.}

\item{...}{Arguments passed to the function implementation.}
}
\value{
Numeric matrix (Nx2) of xy-coordinates for centers of each cell.
}
\description{
Calculate centers of \code{cells} (index 1 to 33) given a xy-point \code{p1}
moving at velocity \code{v1} with angle \code{a1}.
}
