% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{caUtility_rcpp}
\alias{caUtility_rcpp}
\title{Current-angle Utility}
\usage{
caUtility_rcpp(aCA, bCA, bCAlr)
}
\arguments{
\item{aCA}{Numeric scalar power parameter.}

\item{bCA}{Numeric scalar scale parameter.}

\item{bCAlr}{Numeric scalar scale parameter.}
}
\value{
Numeric vector of utilities for each cell.
}
\description{
Current-angle Utility
}
