% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{caUtility_rcpp}
\alias{caUtility_rcpp}
\title{Compute caUtility}
\usage{
caUtility_rcpp(aCA, bCA, bCAlr)
}
\arguments{
\item{aCA}{numeric vector}

\item{bCA}{numeric vector}

\item{bCAlr}{numeric vector}
}
\description{
Compute caUtility
}
