% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coneNum_rcpp}
\alias{coneNum_rcpp}
\title{coneNum}
\usage{
coneNum_rcpp(k)
}
\arguments{
\item{k}{Numeric vector going from 1 to 33}
}
\value{
Numeric vector of length equal to 33
}
\description{
Find equivalent cone number from a vector of angles indices
}
