% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_rcpp_states}
\alias{create_rcpp_states}
\title{create_rcpp_states}
\usage{
create_rcpp_states(states)
}
\arguments{
\item{states}{takes a loaded object of the .rda type}
}
\value{
an empty trace list of lists
}
\description{
Helper functions for creating an empty trace container for subsequent testing
}
