% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_rcpp_trace}
\alias{create_rcpp_trace}
\title{create_rcpp_trace}
\usage{
create_rcpp_trace(trace)
}
\arguments{
\item{trace}{takes a loaded object of the .rda type}
}
\value{
a list of lists containing the .rda's elements
}
\description{
Helper functions for filling the list created by create_rcpp_states for subsequent testing
}
