% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/geometry.R,
%   R/geometry_wrappers.R
\name{dist1_rcpp}
\alias{dist1_rcpp}
\alias{dist1_r}
\alias{dist1}
\title{Vector-to-matrix Distance}
\usage{
dist1_rcpp(p1, p2)

dist1_r(p1, p2)

dist1(..., use = "cpp")
}
\arguments{
\item{p1}{Numeric vector of length two (x and y).}

\item{p2}{Numeric matrix with shape Nx2 (x and y).}
}
\value{
Named numeric vector of length equal to the number of rows in \code{p2}.
}
\description{
Compute the Euclidian distance between a two-element vector and a Nx2
matrix with the first column/element containing x- and the second
column/element y-coordinates.
}
