% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_rcpp}
\alias{dist_rcpp}
\title{dist_rcpp}
\usage{
dist_rcpp(p1, p2)
}
\arguments{
\item{p1}{Numeric matrix}

\item{p2}{Numeric matrix}
}
\value{
Named numeric vector of length equal to the number of rows in p1
}
\description{
Compute distance from p1 to p2, both xy column matrices
}
