% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{flUtility_rcpp}
\alias{flUtility_rcpp}
\title{Follow-leader Utility}
\usage{
flUtility_rcpp(aFL, bFL, dFL, leaders, dists)
}
\arguments{
\item{aFL}{Numeric scalar power parameter.}

\item{bFL}{Numeric scalar scale parameter.}

\item{dFL}{Numeric scalar direction parameter.}

\item{leaders}{Named numeric matrix with columns per leader and rows of their normalized angle disagreement and in-group status.}

\item{dists}{Named numeric matrix with rows per leader and columns per cell with distances from each cell to chosen cell.}
}
\value{
Numeric vector of utilities for each cell.
}
\description{
Follow-leader Utility
}
