% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gaUtility_rcpp}
\alias{gaUtility_rcpp}
\title{Goal Angle Utility}
\usage{
gaUtility_rcpp(bGA, aGA, GA, n_rings = 3L)
}
\arguments{
\item{bGA}{Numeric scalar scale parameter.}

\item{aGA}{Numeric scalar power parameter.}

\item{GA}{Numeric vector of angles to next goal.}

\item{n_rings}{Integer number of cell rings for which to calculate utlity.}
}
\value{
Numeric vector of utilities for each cell.
}
\description{
Goal Angle Utility
}
