% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gaUtility_rcpp}
\alias{gaUtility_rcpp}
\title{Compute gaUtility (Goal angle utility)}
\usage{
gaUtility_rcpp(bGA, aGA, GA)
}
\arguments{
\item{bGA}{integer}

\item{aGA}{integer}

\item{GA}{numeric vector}
}
\value{
a numeric vector of length equal to GA's length
}
\description{
Compute gaUtility (Goal angle utility)
}
