% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/geometry.R,
%   R/geometry_wrappers.R
\name{headingAngle_rcpp}
\alias{headingAngle_rcpp}
\alias{headingAngle_r}
\alias{headingAngle}
\title{Heading Angle}
\usage{
headingAngle_rcpp(a2, a1)

headingAngle_r(
  a2,
  a1,
  angles = c(72.5, 50, 32.5, 20, 10, 0, 350, 340, 327.5, 310, 287.5)
)

headingAngle(..., use = "cpp")
}
\arguments{
\item{a2}{Numeric vector of angles between >0 and 360 degrees.}

\item{a1}{Numeric scalar angles between >0 and 360 degrees.}

\item{angles}{Numeric vector indicating the angle of different directions.
Must be of length 11.}

\item{...}{Arguments passed to the function implementation.}

\item{use}{Whether the R or C++ implementation is called.}
}
\value{
Numeric matrix of shape Nx11 where N is the length of \code{a2}.
}
\description{
Compute absolute angular differences between 11 standard angles and a
vector of angles.
}
\details{
Angle \code{a1} is added to the standard angles.
}
