% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Iangle_rcpp}
\alias{Iangle_rcpp}
\title{Iangle}
\usage{
Iangle_rcpp(p1, p2, a1)
}
\arguments{
\item{p1}{Numeric matrix}

\item{p2}{Numeric matrix}

\item{a1}{Numeric vector - angles in degrees between 0 and 360}
}
\value{
Numeric Vector of indices whereby the angle is within border's bins
}
\description{
Which angle cone (1..11, NA means outside of view) is p2 in relative to p1 heading at angel a1
}
