% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{idUtility_rcpp}
\alias{idUtility_rcpp}
\title{Interpersonal Distance Utility}
\usage{
idUtility_rcpp(bID, dID, aID, n, ok, group, ID_)
}
\arguments{
\item{bID}{Numeric scalar scale parameter.}

\item{dID}{Numeric scalar direction parameter.}

\item{aID}{Numeric scalar power parameter.}

\item{n}{Numeric scalar indexing the subject in the state.}

\item{ok}{Logical matrix indicating if cells are blocked.}

\item{group}{Named numeric scalar with group indices for each pedestrian.}

\item{ID_}{Numeric matrix or NULL; if not NULL, a numeric matrix of predicted distances from the subject to other pedestrians in the front.}
}
\value{
Numeric vector of utilities for each cell.
}
\description{
Interpersonal Distance Utility
}
