% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{idUtility_rcpp}
\alias{idUtility_rcpp}
\title{Compute IDUtility}
\usage{
idUtility_rcpp(bID, dID, aID, n, ok, group, ID_ = NULL)
}
\arguments{
\item{bID}{numeric vector}

\item{dID}{numeric vector}

\item{aID}{numeric vector}

\item{n}{numeric vector}

\item{ok}{logical matrix}

\item{group}{named numeric vector}

\item{ID}{named numeric matrix or NULL by default}
}
\value{
a numeric vector of 33 elements
}
\description{
Compute IDUtility
}
\examples{
idUtility_rcpp(bID = 1.3, aID = 2, dID = 2, n = 1, ID = matrix(rnorm((33*5), mean = 4, sd =1), ncol = 33, nrow = 5, dimnames = list(c('A_1','a_1','b_1','B_1','C_1'))), ok = matrix(data = TRUE, nrow = 11, ncol = 3), group = c('A_1' = 1, 'a_1' = 2, 'b_1' = 3, 'B_1' = 4, 'C_1' = 5, 'c_1' = 6))
idUtility_rcpp(bID = 0, aID = 2, dID = 2, n = 1, ID = NULL, ok = matrix(data = TRUE, nrow = 11, ncol = 3), group = c('A_1' = 1, 'a_1' = 2, 'b_1' = 3, 'B_1' = 4, 'C_1' = 5, 'c_1' = 6))
idUtility_rcpp(bID = 1.3, aID = 2, dID = 2, n = 1, ID = NULL, ok = matrix(data = TRUE, nrow = 11, ncol = 3), group = c('A_1' = 1, 'a_1' = 2, 'b_1' = 3, 'B_1' = 4, 'C_1' = 5, 'c_1' = 6))
}
