% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{idUtility_rcpp}
\alias{idUtility_rcpp}
\title{Compute IDUtility}
\usage{
idUtility_rcpp(bID, dID, aID, n, ok, group, ID_ = NULL)
}
\arguments{
\item{bID}{numeric vector}

\item{dID}{numeric vector}

\item{aID}{numeric vector}

\item{n}{numeric vector}

\item{ok}{logical matrix}

\item{group}{named numeric vector}

\item{ID_}{is an empty matrix of the type NULL}

\item{ID}{named numeric matrix}
}
\value{
a numeric vector of 33 elements
}
\description{
Compute IDUtility
}
