% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{like_state}
\alias{like_state}
\title{State Log-likelihood}
\usage{
like_state(
  state,
  p,
  nests,
  alpha,
  cell_nest,
  elements = "iterations",
  min_like = 1e-10
)
}
\arguments{
\item{state}{List of lists with subject data.}

\item{p}{Numeric matrix with subject parameters for each subject.}

\item{nests}{List of vectors with utility indices.}

\item{alpha}{List of vectors with alpha values.}

\item{cell_nest}{Numeric matrix with nest indices for each cell.}

\item{elements}{Character string indicating whether to return a trace with
iterations or subjects as elements. Can be either \code{"iterations"} or
\code{"subjects"}.}

\item{min_like}{Numeric scalar minimum likelihood return value.}
}
\value{
Numeric vector of state observaton log-likelihoods.
}
\description{
Calculate the log-likelihood of observations for a state as the sum of observation log-likelihoods.
}
