% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minAngle.R
\name{minAngle}
\alias{minAngle}
\title{minAngle}
\usage{
minAngle(a1, a2)
}
\arguments{
\item{a2}{Numeric vector}

\item{a1_double}{scalar vector}
}
\value{
Numeric vector of length equal to a1
}
\description{
Shortest absolute angle between a1 and a2
}
