% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/geometry.R,
%   R/geometry_wrappers.R
\name{minAngle_rcpp}
\alias{minAngle_rcpp}
\alias{minAngle_r}
\alias{minAngle}
\title{Minimum Angle}
\usage{
minAngle_rcpp(a1_double, a2)

minAngle_r(a1, a2)

minAngle(..., use = "cpp")
}
\arguments{
\item{a2}{Numeric vector of angles in between $>0$ and 360 degrees.}

\item{a1, a1_double}{Numeric scalar angle in between $>0$ and 360 degrees.}

\item{...}{Arguments passed to the function implementation.}

\item{use}{Whether the R or C++ implementation is called.}
}
\value{
Numeric vector of length equal to \code{a2} with angles in between
$>0$ and 360 degrees.
}
\description{
Compute smallest absolute angles between a scalar angle and an angle
vector.
}
