% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{psUtility_rcpp}
\alias{psUtility_rcpp}
\title{Preferred Speed Utility}
\usage{
psUtility_rcpp(aPS, bPS, sPref, sSlow, v, d)
}
\arguments{
\item{aPS}{Numeric scalar power parameter.}

\item{bPS}{Numeric scalar scale parameter.}

\item{sPref}{Numeric scalar preference parameter.}

\item{sSlow}{Numeric scalar slowness parameter.}

\item{v}{Numeric scalar current speed.}

\item{d}{Numeric scalar distance to next goal.}
}
\value{
Numeric vector of utilities for each cell.
}
\description{
Preferred Speed Utility
}
