% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/geometry.R,
%   R/geometry_wrappers.R
\name{ringNum_rcpp}
\alias{ringNum_rcpp}
\alias{ringNum_r}
\alias{ringNum}
\title{Ring Number}
\usage{
ringNum_rcpp(k)

ringNum_r(k)

ringNum(..., use = "cpp")
}
\arguments{
\item{k}{Numeric vector between 1 and 33.}
}
\value{
Numeric vector of length equal to k with ring numbers between
1 and 3.
}
\description{
Ring Number
}
