% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scaleVel_rcpp}
\alias{scaleVel_rcpp}
\title{scaleVel}
\usage{
scaleVel_rcpp(v, tStep)
}
\arguments{
\item{v}{Numeric vector}

\item{tStep}{double}
}
\value{
Numeric vector of scaled velocity of same length of v
}
\description{
Scale velocity by time step (tStep)
}
