% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/geometry.R,
%   R/geometry_wrappers.R
\name{scaleVel_rcpp}
\alias{scaleVel_rcpp}
\alias{scaleVel_r}
\alias{scaleVel}
\title{Scaled Velocity}
\usage{
scaleVel_rcpp(v, tStep = 0.5)

scaleVel_r(v, tStep = 0.5)

scaleVel(..., use = "cpp")
}
\arguments{
\item{v}{Numeric vector with velocities.}

\item{tStep}{Numeric scalar scaling factor.}
}
\value{
Numeric vector with scaled velocities of same length as \code{v}.
}
\description{
Scale velocities by a factor.
}
