% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/see.R, R/see_wrappers.R
\name{seesCurrentGoal_rcpp}
\alias{seesCurrentGoal_rcpp}
\alias{seesCurrentGoal_r}
\alias{seesCurrentGoal}
\title{Current Goal in Sight}
\usage{
seesCurrentGoal_rcpp(n, state, objects, offset = 0L)

seesCurrentGoal_r(n, state, objects, offset = 0)

seesCurrentGoal(..., use = "cpp")
}
\arguments{
\item{n}{Integer scalar subject index.}

\item{state}{List of list with state data.}

\item{objects}{List containing a list for each object. An object has
two length-two numeric vectors of x- and y-coordinates.}

\item{offset}{Integer scalar offset to be added to the current goal index
(default 0).}

\item{...}{Arguments passed to the function implementation.}

\item{use}{Whether the R or C++ implementation is called.}
}
\value{
\code{TRUE} if the goal is in sight, \code{FALSE} otherwise.
}
\description{
Checks whether the current goal in a \code{state} can be seen by
subject \code{n}, or if it is occluded by \code{objects}.
}
\examples{
objects = list(
  list(x = c(0.5, 0.5), y = c(0.5, 0.5))
)

state = list(
  p = matrix(c(0, 0), 1, 2),
  P = list(
    matrix(c(1, 1), 1, 2)
  )
)

attr(state$P[[1]], "i") = 1

seesCurrentGoal_rcpp(1, state, objects)
# FALSE

}
