% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{utility}
\alias{utility}
\title{Total Utility of Cells}
\usage{
utility(p, n, v, d, ba_, ga, id_, fl_, wb_, ok, group)
}
\arguments{
\item{p}{Numeric vector of subject parameters.}

\item{n}{Integer scalar indexing the subject in the state.}

\item{v}{Numeric scalar indicating the current speed.}

\item{d}{Numeric scalar indicating the distance to next goal.}

\item{ba_}{NULL or numeric vector of distances from each cell to closest pedestrian.}

\item{ga}{Numeric vector of angles to next goal.}

\item{id_}{NULL or numeric matrix of predicted distances from the subject to other pedestrians in the front.}

\item{fl_}{NULL or list of numeric matrices:
\describe{
  \item{leaders}{Numeric matrix of buddy positions and angles.}
  \item{dists}{Matrix with rows per leader and columns per cell with distances from each cell to chosen cell.}
}}

\item{wb_}{NULL or list of numeric matrices:
\describe{
  \item{buddies}{Matrix with columns per leader and rows of their normalized angle disagreement and in-group status. # needs rewrite}
  \item{dists}{Numeric matrix of distances from cells' centers to closest buddy. # needs rewrite}
}}

\item{ok}{Logical matrix indicating if cells are blocked.}

\item{group}{Integer vector with group indices for each pedestrian.}
}
\value{
Numeric vector with total utility for each cell.
}
\description{
Total Utility of Cells
}
